Moved the first_truthy function to a different file to make it more organized, And so other files can also use the first_truthy function
