+++ src/utils/first_truthy.ts
@@ -1,0 +1,20 @@
+/**
+ * Race multiple async sources and return the first truthy result.
+ * @param sources An array of functions that return Promises of T
+ * @returns The first truthy T, or throws an AggregateError if none do.
+ */
+async function firstTruthy<T>(sources: Array<() => Promise<T>>): Promise<T> {
+    // Wrap each source so that falsy results become rejections
+    const wrapped = sources.map((fn) =>
+        fn().then((value) => {
+            if (value) return value;
+            // convert falsy to rejection so Promise.any will skip it
+            return Promise.reject(new Error("Falsy value"));
+        })
+    );
+
+    // Promise.any returns the first fulfilled promise; if all reject, it rejects
+    return Promise.any(wrapped);
+}
+
+export = firstTruthy;
